requirements.txt
fastapi==0.114.0
uvicorn==0.23.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.2.0
python-dotenv==1.1.1
